requirements.txt
absl-py==2.3.1
contourpy==1.3.3
cycler==0.12.1
drake==1.43.0
etils==1.13.0
fonttools==4.59.0
fsspec==2025.7.0
gitpython==3.1.45
glfw==2.9.0
importlib_resources==6.5.2
jax==0.7.0
jax-cuda12-pjrt==0.7.0
jax-cuda12-plugin==0.7.0
jaxlib==0.7.0
kiwisolver==1.4.8
matplotlib==3.10.3
ml_dtypes==0.5.1
mujoco==3.3.4
mujoco-mjx==3.3.4
numpy==2.3.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.11.0.98
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.27.6
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.9
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
pydot==4.0.1
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.16.0
six==1.17.0
trimesh==4.7.1
typing_extensions==4.14.1
zipp==3.23.0
black==25.1.0
pylint==3.3.7
